Fix test regressions from Decimal refactor and clean up stray file
Fixes a double-negation bug in test_bills.py that silently defeated the
negative-amount validation test, updates test_scenario_workspace_cli.py
to use the get_scenario_workspace() accessor instead of the removed
module-level global, and removes a stray duplicate rcapischemasanalytics.py.

+++ src/api/schemas/budgets.py
@@ -1,5 +1,6 @@
 """Pydantic schemas for budget API requests and responses."""
 
+from decimal import Decimal
 from pydantic import BaseModel, ConfigDict, Field
 
 from src.financial.shared.categories import ExpenseCategory
@@ -9,13 +10,13 @@ class BudgetCreateRequest(BaseModel):
     """Request body for creating or replacing a category budget."""
 
     category: ExpenseCategory
-    limit: float = Field(gt=0)
+    limit: Decimal = Field(gt=0)
 
 
 class BudgetUpdateRequest(BaseModel):
     """Request body for updating a category budget."""
 
-    limit: float = Field(gt=0)
+    limit: Decimal = Field(gt=0)
 
 
 class BudgetResponse(BaseModel):
