+++ src/financial/workflows/account_cash_flow.py
@@ -1,9 +1,10 @@
 from src.financial.accounts.models import Account
+from decimal import Decimal
 
 
 def apply_cash_flow_to_account(
     account: Account,
-    net_cash_flow: float,
+    net_cash_flow: Decimal,
 ) -> Account:
     """Apply net cash flow to an account balance."""
     account.balance += net_cash_flow
